fix: 修复导致一个日志分析无法使用的问题

+++ app/Http/Controllers/BaseController.php
@@ -147,7 +147,7 @@ class BaseController extends Controller
         $lines = array();
         while (count($lines) <= $n) {
             try {
-                fseek($fp, -$pos, SEEK_END);
+                fseek($fp, (int)(-$pos), SEEK_END);
             } catch (Exception $e) {
                 fseek(0);
                 break;
